# HangmanSingle/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Hangman Game (Single User)",
        "description": "A simple API for a Single user Hangman Online/Internet Game",
        "termsOfService": "http://helloreverb.com/terms/",
        "contact": {
            "name": "Hangman API team",
            "email": "ajay.todimala@gmail.com",
            "url": "http://swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/hangman": {
            "post": {
                "description": "Creates a new Hangman Game.",
                "operationId": "createGame",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "game",
                        "in": "body",
                        "description": "Create a new Hangman Game",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/newGame"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "new game response",
                        "schema": {
                            "$ref": "#/definitions/game"
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/errorModel"
                        }
                    }
                }
            }
        },
        "/hangman/{id}": {
            "get": {
                "description": "Returns a Game info for the UI to render the Game to the online player",
                "operationId": "getGameInfoById",
                "produces": [
                    "application/json",
                    "application/xml",
                    "text/xml",
                    "text/html"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of game to fetch",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "game info",
                        "schema": {
                            "$ref": "#/definitions/game"
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/errorModel"
                        }
                    }
                }
            },
            "delete": {
                "description": "deletes a game by id",
                "operationId": "deleteGame",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of game to close",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Game closed"
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/errorModel"
                        }
                    }
                }
            }
        },
        "/hangman/playGame/{id}": {
            "put": {
                "summary": "Play game by guessing a letter",
                "description": "This can only be done by the logged in user.",
                "operationId": "playGame",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "ID of game to play",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Updated Game object",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/game"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid game id supplied"
                    },
                    "404": {
                        "description": "Game not found"
                    }
                }
            }
        },
        "/hangman/guessWord/{id}": {
            "put": {
                "summary": "Play game by guessing a secret word or phrase",
                "description": "This can only be done by the logged in user.",
                "operationId": "guessWord",
                "produces": [
                    "application/json",
                    "application/xml"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "description": "ID of game to play",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Updated Game object",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/game"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid game id supplied"
                    },
                    "404": {
                        "description": "Game not found"
                    }
                }
            }
        }
    },
    "definitions": {
        "game": {
            "type": "object",
            "required": [
                "game_id",
                "sceret_word",
                "guess_word",
                "max_plays",
                "used_plays"
            ],
            "properties": {
                "game_id": {
                    "type": "integer",
                    "format": "int64"
                },
                "sceret_word": {
                    "type": "string"
                },
                "guess_word": {
                    "type": "string"
                },
                "max_plays": {
                    "type": "integer",
                    "format": "int"
                },
                "used_plays": {
                    "type": "integer",
                    "format": "int"
                }
            }
        },
        "newGame": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "tag": {
                    "type": "string"
                }
            }
        },
        "errorModel": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}